Moved factory functions from abstract producer/consumer class to  generic producer/consumer imp
