+++ modules/libInfrastructure/src/kafka_generic_consumer.ts
@@ -65,6 +65,20 @@ export class KafkaGenericConsumer extends MessageConsumer {
     this._logger.info('instance created')
   }
 
+  static Create<tOptions>(options: tOptions, logger: ILogger): MessageConsumer {
+    const consumer = Reflect.construct(this, arguments)
+
+    consumer.on('error', (err: Error): void => {
+      logger.error(`event::error - ${JSON.stringify(err)}`)
+    })
+
+    // consumer.on('commit', (msgMetaData:any) => {
+    //   logger.info(`event::commit - ${JSON.stringify(msgMetaData)}`)
+    // })
+
+    return consumer
+  }
+
   async destroy (forceCommit: boolean = false): Promise<void> {
     return await new Promise((resolve, reject) => {
       if (this._consumerGroup != null) {
